refactor: move quickpay types to repo
