+++ app/src/main/java/to/bitkit/data/CacheStore.kt
@@ -17,6 +17,7 @@ import to.bitkit.models.BalanceState
 import to.bitkit.models.FxRate
 import to.bitkit.models.NewTransactionSheetDetails
 import to.bitkit.models.WalletScope
+import to.bitkit.repositories.QuickPaySpendReservation
 import to.bitkit.utils.Logger
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -178,14 +179,3 @@ data class AppCacheData(
 
     fun invalidateReceiveOnchainAddress() = copy(bip21 = "", onchainAddress = "")
 }
-
-@Serializable
-data class QuickPaySpendReservation(
-    val amountCents: Long,
-    val dayKey: String,
-)
-
-data class QuickPayDaySpend(
-    val dayKey: String,
-    val spentCents: Long,
-)
