+++ app/src/main/java/to/bitkit/repositories/QuickPayRepo.kt
@@ -5,10 +5,9 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
+import kotlinx.serialization.Serializable
 import to.bitkit.data.AppCacheData
 import to.bitkit.data.CacheStore
-import to.bitkit.data.QuickPayDaySpend
-import to.bitkit.data.QuickPaySpendReservation
 import to.bitkit.data.SettingsStore
 import to.bitkit.di.IoDispatcher
 import to.bitkit.ext.runSuspendCatching
@@ -151,6 +150,17 @@ class QuickPayRepo @Inject constructor(
     private fun currentDayKey(): String =
         clock.now().toLocalDateTime(TimeZone.currentSystemDefault()).date.toString()
 }
+
+@Serializable
+data class QuickPaySpendReservation(
+    val amountCents: Long,
+    val dayKey: String,
+)
+
+private data class QuickPayDaySpend(
+    val dayKey: String,
+    val spentCents: Long,
+)
 
 private fun quickPayCapCents(thresholdUsd: Int, multiplier: Int): Long =
     thresholdUsd.toLong() * 100L * multiplier.toLong()
